rearrange swagger to app.ts file
