+++ backend/src/app.ts
@@ -1,3 +1,5 @@
+import swaggerUI from 'swagger-ui-express';
+import swaggerJsDoc from 'swagger-jsdoc';
 import bodyParser from 'body-parser';
 import express, { Request, Response, Express } from 'express';
 import dotenv from 'dotenv';
@@ -24,15 +26,43 @@ const initApp = async (): Promise<Express> => {
     next();
   });
 
+  const swaggerOptions: swaggerJsDoc.Options = {
+    definition: {
+      openapi: '3.0.0',
+      info: {
+        title: 'TGMU Backend API',
+        version: '1.0.0',
+        description: 'TGMU REST API for serving any app related requests including JWT authentication',
+      },
+      servers: [{ url: 'http://localhost:8000' }],
+      components: {
+        securitySchemes: {
+          bearerAuth: {
+            type: 'http',
+            scheme: 'bearer',
+            bearerFormat: 'JWT',
+          },
+        },
+      },
+    },
+    apis: ['./src/routes/*.ts'],
+  };
+
+  logger.debug('Initializing Swagger...');
+  const specs = swaggerJsDoc(swaggerOptions);
+  app.use('/docs', swaggerUI.serve, swaggerUI.setup(specs));
+  logger.debug('Successfully Initialized Swagger at /docs');
+
   app.get('/', (req: Request, res: Response) => {
     res.send('Welcome to The Ger Movie Universe API');
   });
 
   app.use('/auth', authRoute);
-
-  app.use('/users', authMiddleware, userRoute);
-  app.use('/movies', authMiddleware, movieRoute);
   app.use('/public', express.static('public'));
+
+  app.use(authMiddleware);
+  app.use('/users', userRoute);
+  app.use('/movies', movieRoute);
 
   logger.debug('calling init DB');
   await initDB();
